requirements.txt
absl-py==2.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
deepface==0.0.98
dnspython==2.8.0
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.128.6
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
gast==0.7.0
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.78.0
gunicorn==25.0.3
h11==0.16.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.12.1
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
narwhals==2.16.0
numpy==2.2.6
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pillow==12.1.0
protobuf==6.33.5
pyarrow==23.0.0
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pymongo==4.16.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
retina-face==0.0.17
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
starlette==0.52.1
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
Werkzeug==3.1.5
wrapt==2.1.1
